+++ app/models/account/holding_calculator.rb
@@ -1,23 +1,24 @@
-class Account::ReversePortfolioCalculator
+class Account::HoldingCalculator
   def initialize(account)
     @account = account
   end
 
-  def trades 
+  def trades
     account.entries.includes(:entryable).account_trades.to_a
   end
 
-  def portfolio_start_date 
+  def portfolio_start_date
     trades.first ? trades.first.date - 1.day : Date.current
   end
 
   def calculate(reverse: false)
-    reverse ? reverse_holdings : forward_holdings
+    calculated_holdings = reverse ? reverse_holdings : forward_holdings
+    gapfill_holdings(calculated_holdings)
   end
 
   def reverse_holdings
     current_holding_quantities = load_current_holding_quantities
-    prior_holding_quantities = {} 
+    prior_holding_quantities = {}
 
     holdings = []
 
@@ -37,11 +38,11 @@ class Account::ReversePortfolioCalculator
           price: price.price,
           currency: price.currency,
           amount: qty * price.price
-        )
+        ) if price.present?
       end
 
       current_holding_quantities = prior_holding_quantities
-    end 
+    end
 
     holdings
   end
@@ -72,12 +73,45 @@ class Account::ReversePortfolioCalculator
       end
 
       prior_holding_quantities = current_holding_quantities
-    end 
+    end
 
     holdings
   end
 
-  private 
+  def gapfill_holdings(holdings)
+    filled_holdings = []
+
+    holdings.group_by { |h| h.security_id }.each do |security_id, security_holdings|
+      # Skip if no holdings exist for this security
+      next if security_holdings.empty?
+
+      sorted = security_holdings.sort_by(&:date)
+      previous_holding = sorted.first
+
+      sorted.first.date.upto(Date.current) do |date|
+        holding = security_holdings.find { |h| h.date == date }
+
+        if holding
+          filled_holdings << holding
+          previous_holding = holding
+        else
+          # Create a new holding based on the previous day's data
+          filled_holdings << account.holdings.build(
+            security: previous_holding.security,
+            date: date,
+            qty: previous_holding.qty,
+            price: previous_holding.price,
+            currency: previous_holding.currency,
+            amount: previous_holding.amount
+          )
+        end
+      end
+    end
+
+    filled_holdings
+  end
+
+  private
     attr_reader :account
 
     def calculate_quantities(holding_quantities, today_trades, inverse: false)
