+++ app/views/account/transactions/_transaction.html.erb
@@ -34,7 +34,7 @@
   </div>
 
   <% if entry.transfer.present? %>
-    <% unless show_balance %>
+    <% unless balance_trend %>
       <div class="col-span-2"></div>
     <% end %>
