+++ config/locales/defaults/et.yml
@@ -13,7 +13,7 @@ et:
     - E
     - T
     - K
-    - N
+    - "N"
     - R
     - L
     abbr_month_names:
