--- config/locales/views/account/cashes/en.yml
@@ -1,8 +1,0 @@
----
-en:
-  account:
-    cashes:
-      index:
-        cash: Cash
-        name: Name
-        value: Total Balance
