+++ config/locales/views/account/holdings/en.yml
@@ -2,13 +2,13 @@
 en:
   account:
     holdings:
+      cash:
+        brokerage_cash: Brokerage cash
       destroy:
         success: Holding deleted
       holding:
         per_share: per share
         shares: "%{qty} shares"
-      cash:
-        brokerage_cash: Brokerage cash
       index:
         cost: cost
         holdings: Holdings
