+++ config/locales/views/accounts/en.yml
@@ -6,6 +6,8 @@ en:
       troubleshoot: Troubleshoot
     account_list:
       new_account: New %{type}
+    chart:
+      no_change: no change
     create:
       success: "%{type} account created"
     destroy:
@@ -31,8 +33,6 @@ en:
         manual_entry: Enter account balance
         title: How would you like to add it?
       title: What would you like to add?
-    chart:
-      no_change: no change
     show:
       chart:
         balance: Balance
