+++ test/system/transactions_test.rb
@@ -182,7 +182,7 @@ class TransactionsTest < ApplicationSystemTestCase
     investment_account = accounts(:investment)
     investment_account.entries.create!(name: "Investment account", date: Date.current, amount: 1000, currency: "USD", entryable: Account::Transaction.new)
     transfer_date = Date.current
-    visit account_url(investment_account)
+    visit account_url(investment_account, tab: "activity")
     within "[data-testid='activity-menu']" do
       click_on "New"
       click_on "New transaction"
